NCL-7199 Refactor ScmRepositoryLink and ScmRepositoryName
